integrated sidekiq to mailer and export feature
