+++ Gemfile
@@ -2,23 +2,22 @@ source "https://rubygems.org"
 
 gem "rails", "5.2.3"
 
+gem "active_model_serializers", "~> 0.10.0"
 gem "annotate"
 gem "bootstrap-sass"
 gem "devise"
 gem "draper"
 gem "haml-rails"
 gem "jquery-rails"
+gem "redis"
+gem "redis-namespace"
+gem "redis-rack-cache"
+gem "redis-rails"
 gem "sass-rails"
-gem 'sqlite3'
+gem "sidekiq"
+gem "sinatra", github: "sinatra/sinatra"
+gem "sqlite3"
 gem "uglifier"
-gem 'sidekiq'
-gem 'sinatra', github: 'sinatra/sinatra'
-gem 'redis'
-gem 'redis-namespace'
-gem 'redis-rails'
-gem 'redis-rack-cache'
-gem 'active_model_serializers', '~> 0.10.0'
-
 
 # Reduces boot times through caching; required in config/boot.rb
 gem "bootsnap", ">= 1.1.0", require: false
@@ -36,11 +35,11 @@ end
 
 group :development, :test do
   gem "capybara"
-  gem 'faker'
   gem "factory_bot_rails"
+  gem "faker"
   gem "listen"
-  gem "rspec-rails"
   gem "pry-byebug"
+  gem "rspec-rails"
   gem "rubocop"
   gem "simplecov"
 end
