+++ app/controllers/api/base_controller.rb
@@ -1,0 +1,1 @@
+class Api::BaseController < ActionController::Base; end
