+++ app/controllers/api/v2/base_controller.rb
@@ -1,0 +1,2 @@
+class Api::V2::BaseController < Api::BaseController
+end
