+++ app/mailers/movie_info_mailer.rb
@@ -1,7 +1,9 @@
 class MovieInfoMailer < ApplicationMailer
   def send_info(user, movie)
     @movie = movie
-    sleep(3) # this emulates long email sending, do not remove
+    user_json = user.as_json
+    movie_json = movie.as_json
+    MovieInfoWorker.perform_async(user_json,movie_json)
     mail(to: user.email, subject: "Info about movie #{movie.title}")
   end
 end
