--- app/queries/number_of_movies_query.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module NumberOfMoviesQuery
-  def initialize(genre_id)
-    Genre.find(params[genre_id]).movies.count
-  end
-end
