+++ app/workers/movie_export_worker.rb
@@ -1,0 +1,26 @@
+class MovieExportWorker
+  include Sidekiq::Worker
+  sidekiq_options retry: false
+  require 'csv'
+
+  def perform(user,file_path)
+    CSV.open(file_path, "wb", csv_options) do |csv|
+      Movie.all.each do |movie|
+        csv << [movie.title, movie.description]
+      end
+    end
+    sleep(5) # this emulates long export, do not remove
+    MovieExportMailer.send_file(user, file_path).deliver_now
+  end
+
+  private
+
+  def csv_options
+    {
+        write_headers: true,
+        headers: %w[title description],
+        col_sep: ";"
+    }
+  end
+
+end
