--- app/workers/movie_worker.rb
@@ -1,8 +1,0 @@
-class MovieWorker
-  include Sidekiq::Worker
-  sidekiq_options retry: false
-
-  def perform(movie)
-  end
-
-end
