+++ spec/workers/hard_worker_spec.rb
@@ -1,4 +1,4 @@
 require 'rails_helper'
-RSpec.describe MovieWorker, type: :worker do
+RSpec.describe MovieInfoWorker, type: :worker do
   pending "add some examples to (or delete) #{__FILE__}"
 end
